--- a/crates/data_preparation/src/dataloader/iterator/inmemory.rs
+++ b/crates/data_preparation/src/dataloader/iterator/inmemory.rs
@@ -62,7 +62,6 @@ where
             LoaderType::InMemory {
                 batch_sampler,
                 worker_manager,
-                uses_internal_shuffle: _,
             } => {
                 let batch_indices = batch_sampler.iter(sampler_epoch);