--- a/crates/data_preparation/src/dataset.rs
+++ b/crates/data_preparation/src/dataset.rs
@@ -1,5 +1,5 @@
 use crate::sample::Sample;
-use crate::transform::Transform;
+use crate::transforms::Transform;
 
 use anyhow::{anyhow, Result};
 use std::collections::HashMap;