--- a/crates/data_preparation/src/readers/mod.rs
+++ b/crates/data_preparation/src/readers/mod.rs
@@ -1,0 +1,3 @@
+pub mod txt; 
+
+pub use txt::TxtSource;