## Overview
This PR implements the `DataLoader` module for efficient data loading in machine learning pipelines. The implementation supports both in-memory and streaming datasets with single-threaded and multi-threaded execution modes.

## Features

### 1. Core Components
- **DataLoader**: Main orchestrator that coordinates datasets, samplers, and workers
- **DataLoaderConfig**: Configuration with builder pattern for easy setup
- **Worker Pool**: Thread pool for parallel data loading
- **Iterator System**: Flexible iteration with 6 specialized variants (see Iterator Variants below)

### 2. Dataset Support
- **InMemoryDataset**: Random access with index-based sampling
  - **Auto-sampling:** When no sampler is provided: 
      - `shuffle = true` -> automatically uses `RandomSampler`
      - `shuffle = false` -> automatically uses `SequentialSampler`
  - **Custom sampling:** Users can provide their own sampler (Random, Sequential, Subset, Weighted) 
  - **Custom batch sampling:** Users can also provide their own pre-batch sampler (e.g., `BatchBucketSampler`)
  
- **IterableDataset**: Streaming with shard-based distribution
  - Round-robin shard assignment to workers
  - No duplicate processing across workers
  - Warning system for idle workers: Warns when `num_worker > num_sources`

#### Design Rationale: Separate Constructors
Due to fundamental differences between in-memory (indexable) and streaming (non-indexable) datasets, we implement separate code paths with 6 different constructors to ensure:

- **Type safety:** Invalid combinations caught at compile time
- **Clear contracts:** e.g., batch samplers cannot be used with `batch_size`/`shuffle`/`drop_last`

A unified interface would require runtime checks and degrade performance. See `src/dataloader/loader.rs` for the complete constructor matrix. 

### 3. Worker Modes
- **Single-threaded**: Simple, low-memory operation
- **Fresh workers**: Created per epoch, good for simple transforms
- **Persistent workers**: Reused across epochs, reduces overhead

### 4. Determinism & Reproducibility
The DataLoader ensures reproducible training through coordinated seed handling:
- **Sampler seed:** Controls data ordering across epochs
- **Worker seed:**  `base_seed + (epoch << 32) + worker_id` ensures:
    - Same worker always processes same batches (reproducible assignment)
    - Different random augmentations per worker (no duplicate transforms)
    - Consistent results across runs with same seed

This solves the [classic PyTorch worker seed issue](https://gist.github.com/NicolasHug/96a75c2d754ff2a7c52afca2c0b628d4) where workers inherit identical random states.  The `DataLoader` constructor validates seed consistency and errors out if the sampler and config seeds do not match.

### 5. Performance Features
- **Prefetch pipeline:** Configurable prefetch_factor keeps GPU fed
- **Backpressure:** Bounded channels prevent memory explosion
- **Zero-copy sharing:** Dataset wrapped in Arc for efficient worker access
- **Lazy loading:** Transforms applied on-demand, not pre-cached. 

To support lazy loading, we added `get_sample(index)` to `InMemoryDataset`:
```rust
// Before: Each worker caches entire transformed dataset
worker → dataset.iter() → transform all → cache → serve batches ❌

// After: On-demand transformation
worker → receive indices → dataset.get_sample(idx) → transform → batch ✓
```
Memory impact: O(num_workers × dataset_size) → O(num_workers × batch_size)

## Architecture Highlights

### Iterator Variants
The implementation provides 6 specialized iterator variants optimized for different scenarios:
- `InMemorySingle` / `IterableSingle`: Single-threaded baseline
- `InMemoryMulti` / `IterableMulti`: Multi-threaded with fresh workers
- `InMemoryPersistent` / `IterablePersistent`: Multi-threaded with worker reuse

### Worker Communication
- **InMemory**: Task-based with deterministic batch assignment
- **Iterable**: Control channels for epoch synchronization
- Graceful shutdown and timeout handling

## Usage Examples

```rust
// Basic usage with auto-sampling
let config = DataLoaderConfig::builder()
    .batch_size(32)
    .num_workers(4)
    .shuffle(true)
    .seed(42)
    .build();

let dataloader = DataLoader::new(dataset, config)?;

for batch in dataloader.iter()? {
    let batch = batch?;
    // Process batch...
}

// Advanced: Custom sampler with persistent workers
let sampler = WeightedRandomSampler::new(weights, false, None, 42)?;
let config = DataLoaderConfig::builder()
    .batch_size(64)
    .num_workers(8)
    .persistent_workers(true)
    .prefetch_factor(2)
    .build();

let dataloader = DataLoader::new_with_sampler(dataset, sampler, config)?;
```

## Preliminary Benchmark Results

### 1. Throughput Benchmarks

| Workload | 0 workers | 2 workers | 4 workers |
| :--- | :--- | :--- | :--- |
| **InMemory - Light transform** | 2.96ms | 1.72ms (1.7x) | 1.27ms (2.3x) |
| **InMemory - Heavy transform (Tokenizer)** | 2.04s | 1.03s (2.0x) | 533ms (3.8x) |
| **Iterable - I/O streaming** | 738ms | 378ms (2.0x) | 190ms (3.9x) |

### 2. Persistent Workers Performance

| Scenario | Fresh Workers | Persistent Workers | Improvement |
| :--- | :--- | :--- | :--- |
| **Multi-epoch (50 epochs)** | 712ms | 160ms | **4.5x** faster |
| **Heavy transforms** | 200ms | 200ms | No overhead |
| **With initialization cost** | 153ms/epoch | 1.25ms/epoch | **122x** faster |

### Key Findings

- **Near-linear scaling** with worker count for both CPU and I/O bound workloads.
- **Persistent workers** shine in multi-epoch training, eliminating repeated initialization costs.
- **No performance overhead** for persistent workers in single-epoch scenarios.
- **Huge speedup (122x)** when transforms have initialization costs (e.g., loading models).

The benchmarks demonstrate that our `DataLoader` implementation can effectively parallelize work without significant overhead. But of course, more rigorous benchmarking is necessary. 

## Testing
All `DataLoader` tests are within the tests folder. Covers: 
1. Basic functionality for all dataset types and configurations
2. Determinism and seed handling across workers/epochs
3. Worker lifecycle management (fresh vs persistent)
4. Error propagation and edge cases
5. Concurrency correctness (no races, proper backpressure)

## Future Work
1. Performance profiling with `cargo flamegraph` and benchmarks against PyTorch
2. Stress testing on real workloads (ImageNet, large NLP datasets)
3. Documentation improvements and usage guide
4. GPU memory pinning support
5. Async/await for I/O-bound workloads